+++ CHANGELOG.md
@@ -9,3 +9,5 @@
 - Added get balance method to fetch STX balance of an account and test cases
 - Added functionality to sign and send native STX token, fungible tokens
 - Added functionality to sign message
+- Added ckecks to function params
+- Added test cases for sign message and transaction function